requirements.txt
# pythonLibs Requirements
digital-thought-commons>=1.0.0
pyvelociraptor
grpcio
xlsxwriter
cryptography